Fix parallelization start issue with high-latency IO
